+++ src/minerfund.h
@@ -18,7 +18,9 @@ struct Amount;
 struct Params;
 } // namespace Consensus
 
-Amount GetMinerFundAmount(const Amount &coinbaseValue);
+Amount GetMinerFundAmount(const Consensus::Params &params,
+                          const Amount &coinbaseValue,
+                          const CBlockIndex *pprev);
 
 std::unordered_set<CTxDestination, TxDestinationHasher>
 GetMinerFundWhitelist(const Consensus::Params &params);
@@ -28,6 +30,6 @@ GetMinerFundWhitelist(const Consensus::Params &params);
  */
 bool CheckMinerFund(const Consensus::Params &params,
                     const std::vector<CTxOut> &coinbaseTxOut,
-                    const Amount &blockReward);
+                    const Amount &blockReward, const CBlockIndex *pprev);
 
 #endif // BITCOIN_MINERFUND_H