+++ doc/release-notes.md
@@ -4,4 +4,10 @@ Bitcoin ABC version 0.28.0 is now available from:
 
   <https://download.bitcoinabc.org/0.28.0/>
 
-This is a maintenance release with no user-visible change.
+Network upgrade
+---------------
+
+At the MTP time of `1700049600` (November 15, 2023 12:00:00 UTC), the following
+changes will become activated:
+ - The miner fund ratio is increased to 32% of the coinbase reward
+